pyproject.toml
[project]
name = "todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==24.1.0",
    "alembic>=1.17.2",
    "amqp==5.3.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "bcrypt==4.0.1",
    "billiard==4.2.3",
    "celery==5.5.3",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "click==8.2.1",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1.2",
    "click-repl==0.3.0",
    "colorama==0.4.6 ; sys_platform == 'win32'",
    "cryptography==45.0.5",
    "dnspython==2.7.0",
    "ecdsa==0.19.1",
    "email-validator==2.2.0",
    "fastapi==0.116.1",
    "fastapi-cli==0.0.8",
    "fastapi-cloud-cli==0.1.5",
    "flower==2.0.1",
    "graphql-core==3.2.7",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "humanize==4.14.0",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jinja2==3.1.6",
    "kombu==5.5.4",
    "lia-web==0.2.3",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "mypy==1.18.2",
    "mypy-extensions==1.1.0",
    "packaging==25.0",
    "passlib==1.7.4",
    "pathspec==0.12.1",
    "pluggy==1.6.0",
    "prometheus-client==0.23.1",
    "prompt-toolkit==3.0.52",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pymysql==1.1.1",
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "redis==5.2.1",
    "rich==14.1.0",
    "rich-toolkit==0.14.9",
    "rignore==0.6.4",
    "rsa==4.9.1",
    "sentry-sdk==2.34.1",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.42",
    "starlette==0.47.2",
    "strawberry-graphql==0.285.0",
    "tornado==6.5.2",
    "typer==0.16.0",
    "types-passlib==1.7.7.20250602",
    "types-pyasn1==0.6.0.20250914",
    "types-python-jose==3.5.0.20250531",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "uvloop==0.21.0",
    "vine==5.1.0",
    "watchfiles==1.1.0",
    "wcwidth==0.2.14",
    "websockets==15.0.1",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = true

# exclude alembic directory from mypy checks
exclude = 'alembic/.*'

[dependency-groups]
dev = [
    "isort>=7.0.0",
]
